Jeremy, qué hiciste

+++ visita-facil-api/src/main/java/com/construsoft/visita_facil_api/domain/SolicitudVisitaDTO.java
@@ -51,8 +51,8 @@ public class SolicitudVisitaDTO {
         this.hora = hora;
     }
 
-    public Long getIdPropiedad() {
-        return (long) Math.toIntExact(idPropiedad);
+    public Integer getIdPropiedad() {
+        return (Integer) Math.toIntExact(idPropiedad);
     }
 
     public void setIdPropiedad(Long idPropiedad) {
